# api/openapi/openapi.yaml
openapi: "3.0.1"
info:
  title: URL Shortener API
  description: URL Shortener API"
  version: 1.0.0
  contact:
    email: alekceeb@gmail.com
components:
  schemas:
    RequestURL:
      type: object
      properties:
        url:
          type: string
          format: url
    ResponseURL:
      type: object
      properties:
        shortURL:
          type: string
          format: url
        statsURL:
          type: string
          format: url
    Stats:
      type: object
      properties:
        shortURL:
          type: string
          format: url
        numRedirects:
          type: integer
          format: int64
tags:
- name: Short URL
  description: Create short URL
- name: Stats
  description: Get stats

paths:
  /:
    post:
      summary: Create short URL from original URL
      tags:
        - Short URL
      operationId: CreateShortURL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestURL"
      responses:
        201:
          description: short url created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseURL"
        400:
          description: bad request
          content: {}
        500:
          description: internal server error
          content: {}
  /{short-url}:
    get:
      summary: Redirect from short URL to original URL
      tags:
        - Short URL
      operationId: RedirectURL
      parameters:
        - name: short-url
          in: path
          description: short URL for redirecting
          required: true
          schema:
            type: string
      responses:
        303:
          description: successful operation
        404:
          description: not found
        500:
          description: internal server error
  /{short-url}/stats:
    get:
      summary: Get stats about redirects
      tags:
        - Stats
      operationId: GetStats
      parameters:
        - name: short-url
          in: path
          description: short URL for getting redirect stats
          required: true
          schema:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"
        404:
          description: not found
        500:
          description: internal server error